Add SendMoney component and integrate with Accounts page; increase debounce delay in DebitRequestModal and Send Money's Company & individual Components.
