+++ src/ui/components/SingleAccount/Individual.tsx
@@ -31,6 +31,7 @@ import {
   Badge,
   Loader,
   ActionIcon,
+  ScrollArea,
 } from "@mantine/core";
 import { TextInput, Select, Button, UnstyledButton } from "@mantine/core";
 import { useForm, UseFormReturnType, zodResolver } from "@mantine/form";
@@ -91,7 +92,7 @@ function Individual({
 
   const [{ bic, iban }] = useDebouncedValue(
     { iban: form.values.destinationIBAN, bic: form.values.destinationBIC },
-    500
+    2000
   );
 
   const handleIbanValidation = async () => {
@@ -146,236 +147,238 @@ function Individual({
     <>
       <TabsPanel value="To Individual">
         <Box mt={20}>
-          <Flex gap={20}>
-            <TextInput
-              classNames={{ input: styles.input, label: styles.label }}
-              flex={1}
-              size="lg"
-              label={
-                <Text fz={14} c="#667085">
-                  First Name
-                </Text>
-              }
-              placeholder="Enter first name"
-              {...form.getInputProps("firstName")}
-              errorProps={{
-                fz: 12,
-              }}
-            />
-
-            <TextInput
-              classNames={{ input: styles.input, label: styles.label }}
-              flex={1}
-              size="lg"
-              label={
-                <Text fz={14} c="#667085">
-                  Last Name
-                </Text>
-              }
-              placeholder="Enter last name"
-              {...form.getInputProps("lastName")}
-              errorProps={{
-                fz: 12,
-              }}
-            />
-          </Flex>
-
-          <Flex gap={20} mt={24}>
-            <TextInput
-              classNames={{ input: styles.input, label: styles.label }}
-              flex={1}
-              size="lg"
-              label={
-                <Text fz={14} c="#667085">
-                  IBAN
-                </Text>
-              }
-              placeholder="Enter IBAN"
-              {...form.getInputProps("destinationIBAN")}
-              errorProps={{
-                fz: 12,
-              }}
-            />
-
-            <TextInput
-              classNames={{ input: styles.input, label: styles.label }}
-              flex={1}
-              size="lg"
-              label={
-                <Text fz={14} c="#667085">
-                  BIC
-                </Text>
-              }
-              placeholder="Enter BIC"
-              {...form.getInputProps("destinationBIC")}
-              errorProps={{
-                fz: 12,
-              }}
-            />
-          </Flex>
-
-          {(processing || validated) && (
-            <Group
-              justify="space-between"
-              bg="#ECFDF3"
-              w="100%"
-              px={20}
-              py={8}
-              my={32}
-            >
-              <Badge
-                fz={14}
-                px={0}
-                c="#12B76A"
-                variant="transparent"
-                fw={600}
-                color="#12B76A"
-                tt="capitalize"
-                rightSection={
-                  validated ? (
-                    <ActionIcon
-                      variant="light"
-                      radius="xl"
-                      color="#12B76A"
-                      size={23}
-                    >
-                      <IconCheck />
-                    </ActionIcon>
-                  ) : null
+          <ScrollArea h="calc(100dvh - 500px)" scrollbarSize={3} pr={20}>
+            <Flex gap={20}>
+              <TextInput
+                classNames={{ input: styles.input, label: styles.label }}
+                flex={1}
+                size="lg"
+                label={
+                  <Text fz={14} c="#667085">
+                    First Name
+                  </Text>
                 }
+                placeholder="Enter first name"
+                {...form.getInputProps("firstName")}
+                errorProps={{
+                  fz: 12,
+                }}
+              />
+
+              <TextInput
+                classNames={{ input: styles.input, label: styles.label }}
+                flex={1}
+                size="lg"
+                label={
+                  <Text fz={14} c="#667085">
+                    Last Name
+                  </Text>
+                }
+                placeholder="Enter last name"
+                {...form.getInputProps("lastName")}
+                errorProps={{
+                  fz: 12,
+                }}
+              />
+            </Flex>
+
+            <Flex gap={20} mt={24}>
+              <TextInput
+                classNames={{ input: styles.input, label: styles.label }}
+                flex={1}
+                size="lg"
+                label={
+                  <Text fz={14} c="#667085">
+                    IBAN
+                  </Text>
+                }
+                placeholder="Enter IBAN"
+                {...form.getInputProps("destinationIBAN")}
+                errorProps={{
+                  fz: 12,
+                }}
+              />
+
+              <TextInput
+                classNames={{ input: styles.input, label: styles.label }}
+                flex={1}
+                size="lg"
+                label={
+                  <Text fz={14} c="#667085">
+                    BIC
+                  </Text>
+                }
+                placeholder="Enter BIC"
+                {...form.getInputProps("destinationBIC")}
+                errorProps={{
+                  fz: 12,
+                }}
+              />
+            </Flex>
+
+            {(processing || validated) && (
+              <Group
+                justify="space-between"
+                bg="#ECFDF3"
+                w="100%"
+                px={20}
+                py={8}
+                my={32}
               >
-                {validated
-                  ? "Information Validated "
-                  : "Verifying Account Details"}
-              </Badge>
+                <Badge
+                  fz={14}
+                  px={0}
+                  c="#12B76A"
+                  variant="transparent"
+                  fw={600}
+                  color="#12B76A"
+                  tt="capitalize"
+                  rightSection={
+                    validated ? (
+                      <ActionIcon
+                        variant="light"
+                        radius="xl"
+                        color="#12B76A"
+                        size={23}
+                      >
+                        <IconCheck />
+                      </ActionIcon>
+                    ) : null
+                  }
+                >
+                  {validated
+                    ? "Information Validated "
+                    : "Verifying Account Details"}
+                </Badge>
 
-              {processing && <Loader type="oval" size={24} color="#12B76A" />}
-            </Group>
-          )}
+                {processing && <Loader type="oval" size={24} color="#12B76A" />}
+              </Group>
+            )}
 
-          {validated && (
-            <>
-              {" "}
-              <Flex gap={20} mt={24}>
-                <TextInput
-                  placeholder="Enter Bank Name"
-                  classNames={{ input: styles.input, label: styles.label }}
-                  flex={1}
-                  size="lg"
-                  label={
-                    <Text fz={14} c="#667085">
-                      Bank
-                    </Text>
-                  }
-                  disabled={disabled}
-                  {...form.getInputProps("destinationBank")}
-                  errorProps={{
-                    fz: 12,
-                  }}
-                />
-              </Flex>
-              <Flex gap={20} mt={24}>
-                <TextInput
-                  classNames={{ input: styles.input, label: styles.label }}
-                  flex={1}
-                  size="lg"
-                  label={
-                    <Text fz={14} c="#667085">
-                      Bank Address
-                    </Text>
-                  }
-                  placeholder="Bank Address"
-                  disabled={disabled}
-                  {...form.getInputProps("bankAddress")}
-                  errorProps={{
-                    fz: 12,
-                  }}
-                />
-              </Flex>
-              <Flex gap={20} mt={24}>
-                <NumberInput
-                  flex={1}
-                  classNames={{ input: styles.input, label: styles.label }}
-                  description={
-                    <Text fz={12}>
-                      {Number(form.values.amount) >
-                      Number(account?.accountBalance)
-                        ? `Insufficient Balance`
-                        : ""}
-                    </Text>
-                  }
-                  styles={{
-                    description: {
-                      color: "var(--prune-warning)",
-                    },
-                    input: {
-                      border:
-                        Number(form.values.amount) >
+            {validated && (
+              <>
+                {" "}
+                <Flex gap={20} mt={24}>
+                  <TextInput
+                    placeholder="Enter Bank Name"
+                    classNames={{ input: styles.input, label: styles.label }}
+                    flex={1}
+                    size="lg"
+                    label={
+                      <Text fz={14} c="#667085">
+                        Bank
+                      </Text>
+                    }
+                    disabled={disabled}
+                    {...form.getInputProps("destinationBank")}
+                    errorProps={{
+                      fz: 12,
+                    }}
+                  />
+                </Flex>
+                <Flex gap={20} mt={24}>
+                  <TextInput
+                    classNames={{ input: styles.input, label: styles.label }}
+                    flex={1}
+                    size="lg"
+                    label={
+                      <Text fz={14} c="#667085">
+                        Bank Address
+                      </Text>
+                    }
+                    placeholder="Bank Address"
+                    disabled={disabled}
+                    {...form.getInputProps("bankAddress")}
+                    errorProps={{
+                      fz: 12,
+                    }}
+                  />
+                </Flex>
+                <Flex gap={20} mt={24}>
+                  <NumberInput
+                    flex={1}
+                    classNames={{ input: styles.input, label: styles.label }}
+                    description={
+                      <Text fz={12}>
+                        {Number(form.values.amount) >
                         Number(account?.accountBalance)
-                          ? "1px solid red"
-                          : "1px solid #eaecf0",
-                    },
-                  }}
-                  label={
-                    <Text fz={14} c="#667085">
-                      Amount
-                    </Text>
-                  }
-                  hideControls
-                  size="lg"
-                  placeholder="Enter amount"
-                  {...form.getInputProps("amount")}
-                  errorProps={{
-                    fz: 12,
-                  }}
-                  // error={account?.accountBalance < form.values.amount}
-                />
-              </Flex>
-              <Flex gap={20} mt={24}>
-                <Select
-                  searchable
-                  placeholder="Select Country"
-                  classNames={{ input: styles.input, label: styles.label }}
-                  flex={1}
-                  label="Country"
-                  data={countries}
-                  disabled={disabled}
-                  {...form.getInputProps("destinationCountry")}
-                />
-              </Flex>
-              <Flex gap={20} mt={24} direction="column">
-                <Text fz={14} c="#667085" m={0} p={0}>
-                  Upload supporting document (Optional)
-                </Text>
-                <DropzoneComponent<typeof sendMoneyIndividualRequest>
-                  style={{ flex: 1 }}
-                  otherForm={form}
-                  formKey="invoice"
-                  uploadedFileUrl={form.values.invoice}
-                  isUser
-                />
-              </Flex>
-              <Flex gap={20} mt={24}>
-                <Textarea
-                  flex={1}
-                  autosize
-                  minRows={3}
-                  size="lg"
-                  classNames={{ input: styles.textarea, label: styles.label }}
-                  label={
-                    <Text fz={14} c="#667085">
-                      Narration
-                    </Text>
-                  }
-                  placeholder="Enter narration"
-                  {...form.getInputProps("narration")}
-                  errorProps={{
-                    fz: 12,
-                  }}
-                />
-              </Flex>
-            </>
-          )}
+                          ? `Insufficient Balance`
+                          : ""}
+                      </Text>
+                    }
+                    styles={{
+                      description: {
+                        color: "var(--prune-warning)",
+                      },
+                      input: {
+                        border:
+                          Number(form.values.amount) >
+                          Number(account?.accountBalance)
+                            ? "1px solid red"
+                            : "1px solid #eaecf0",
+                      },
+                    }}
+                    label={
+                      <Text fz={14} c="#667085">
+                        Amount
+                      </Text>
+                    }
+                    hideControls
+                    size="lg"
+                    placeholder="Enter amount"
+                    {...form.getInputProps("amount")}
+                    errorProps={{
+                      fz: 12,
+                    }}
+                    // error={account?.accountBalance < form.values.amount}
+                  />
+                </Flex>
+                <Flex gap={20} mt={24}>
+                  <Select
+                    searchable
+                    placeholder="Select Country"
+                    classNames={{ input: styles.input, label: styles.label }}
+                    flex={1}
+                    label="Country"
+                    data={countries}
+                    disabled={disabled}
+                    {...form.getInputProps("destinationCountry")}
+                  />
+                </Flex>
+                <Flex gap={20} mt={24} direction="column">
+                  <Text fz={14} c="#667085" m={0} p={0}>
+                    Upload supporting document (Optional)
+                  </Text>
+                  <DropzoneComponent<typeof sendMoneyIndividualRequest>
+                    style={{ flex: 1 }}
+                    otherForm={form}
+                    formKey="invoice"
+                    uploadedFileUrl={form.values.invoice}
+                    isUser
+                  />
+                </Flex>
+                <Flex gap={20} mt={24}>
+                  <Textarea
+                    flex={1}
+                    autosize
+                    minRows={3}
+                    size="lg"
+                    classNames={{ input: styles.textarea, label: styles.label }}
+                    label={
+                      <Text fz={14} c="#667085">
+                        Narration
+                      </Text>
+                    }
+                    placeholder="Enter narration"
+                    {...form.getInputProps("narration")}
+                    errorProps={{
+                      fz: 12,
+                    }}
+                  />
+                </Flex>
+              </>
+            )}
+          </ScrollArea>
 
           <Flex mt={24} justify="flex-end" gap={15}>
             <SecondaryBtn
