+++ src/ui/components/SingleAccount/company.tsx
@@ -29,6 +29,7 @@ import {
   ActionIcon,
   Badge,
   Loader,
+  ScrollArea,
 } from "@mantine/core";
 import { TextInput, Select, Button, UnstyledButton } from "@mantine/core";
 import { useForm, zodResolver } from "@mantine/form";
@@ -85,7 +86,7 @@ function Company({
 
   const [{ bic, iban }] = useDebouncedValue(
     { iban: form2.values.destinationIBAN, bic: form2.values.destinationBIC },
-    500
+    2000
   );
 
   const handlePreviewState = () => {
@@ -132,229 +133,234 @@ function Company({
 
   useEffect(() => {
     if (iban && bic) {
+      console.log({ bic, iban });
       handleIbanValidation();
     }
   }, [bic, iban]);
 
+  console.log({ bic, iban });
+
   return (
     <TabsPanel value="To A Company">
       <Box mt={20}>
-        <Flex gap={20}>
-          <TextInput
-            classNames={{ input: styles.input, label: styles.label }}
-            flex={1}
-            size="lg"
-            label={
-              <Text fz={14} c="#667085">
-                Company Name
-              </Text>
-            }
-            placeholder="Enter company Name"
-            {...form2.getInputProps("companyName")}
-            errorProps={{
-              fz: 12,
-            }}
-          />
-        </Flex>
-        <Flex gap={20} mt={24}>
-          <TextInput
-            classNames={{ input: styles.input, label: styles.label }}
-            flex={1}
-            size="lg"
-            label={
-              <Text fz={14} c="#667085">
-                IBAN
-              </Text>
-            }
-            placeholder="Enter IBAN"
-            {...form2.getInputProps("destinationIBAN")}
-            errorProps={{
-              fz: 12,
-            }}
-          />
-
-          <TextInput
-            classNames={{ input: styles.input, label: styles.label }}
-            flex={1}
-            size="lg"
-            label={
-              <Text fz={14} c="#667085">
-                BIC
-              </Text>
-            }
-            placeholder="Enter BIC"
-            {...form2.getInputProps("destinationBIC")}
-            errorProps={{
-              fz: 12,
-            }}
-          />
-        </Flex>
-
-        {(processing || validated) && (
-          <Group
-            justify="space-between"
-            bg="#ECFDF3"
-            w="100%"
-            px={20}
-            py={8}
-            my={32}
-          >
-            <Badge
-              fz={14}
-              px={0}
-              c="#12B76A"
-              variant="transparent"
-              fw={600}
-              color="#12B76A"
-              tt="capitalize"
-              rightSection={
-                validated ? (
-                  <ActionIcon
-                    variant="light"
-                    radius="xl"
-                    color="#12B76A"
-                    size={23}
-                  >
-                    <IconCheck />
-                  </ActionIcon>
-                ) : null
+        <ScrollArea h="calc(100dvh - 500px)" scrollbarSize={3} pr={20}>
+          <Flex gap={20}>
+            <TextInput
+              classNames={{ input: styles.input, label: styles.label }}
+              flex={1}
+              size="lg"
+              label={
+                <Text fz={14} c="#667085">
+                  Company Name
+                </Text>
               }
+              placeholder="Enter company Name"
+              {...form2.getInputProps("companyName")}
+              errorProps={{
+                fz: 12,
+              }}
+            />
+          </Flex>
+          <Flex gap={20} mt={24}>
+            <TextInput
+              classNames={{ input: styles.input, label: styles.label }}
+              flex={1}
+              size="lg"
+              label={
+                <Text fz={14} c="#667085">
+                  IBAN
+                </Text>
+              }
+              placeholder="Enter IBAN"
+              {...form2.getInputProps("destinationIBAN")}
+              errorProps={{
+                fz: 12,
+              }}
+            />
+
+            <TextInput
+              classNames={{ input: styles.input, label: styles.label }}
+              flex={1}
+              size="lg"
+              label={
+                <Text fz={14} c="#667085">
+                  BIC
+                </Text>
+              }
+              placeholder="Enter BIC"
+              {...form2.getInputProps("destinationBIC")}
+              errorProps={{
+                fz: 12,
+              }}
+            />
+          </Flex>
+
+          {(processing || validated) && (
+            <Group
+              justify="space-between"
+              bg="#ECFDF3"
+              w="100%"
+              px={20}
+              py={8}
+              my={32}
             >
-              {validated
-                ? "Information Validated "
-                : "Verifying Account Details"}
-            </Badge>
-
-            {processing && <Loader type="oval" size={24} color="#12B76A" />}
-          </Group>
-        )}
-
-        {validated && (
-          <>
-            <Flex gap={20} mt={24}>
-              <TextInput
-                placeholder="Enter Bank Name"
-                classNames={{ input: styles.input, label: styles.label }}
-                flex={1}
-                size="lg"
-                label={
-                  <Text fz={14} c="#667085">
-                    Bank
-                  </Text>
+              <Badge
+                fz={14}
+                px={0}
+                c="#12B76A"
+                variant="transparent"
+                fw={600}
+                color="#12B76A"
+                tt="capitalize"
+                rightSection={
+                  validated ? (
+                    <ActionIcon
+                      variant="light"
+                      radius="xl"
+                      color="#12B76A"
+                      size={23}
+                    >
+                      <IconCheck />
+                    </ActionIcon>
+                  ) : null
                 }
-                disabled={disabled}
-                {...form2.getInputProps("destinationBank")}
-                errorProps={{
-                  fz: 12,
-                }}
-              />
-            </Flex>
+              >
+                {validated
+                  ? "Information Validated "
+                  : "Verifying Account Details"}
+              </Badge>
 
-            <Flex gap={20} mt={24}>
-              <TextInput
-                classNames={{ input: styles.input, label: styles.label }}
-                flex={1}
-                size="lg"
-                label={
-                  <Text fz={14} c="#667085">
-                    Bank Address
-                  </Text>
-                }
-                disabled={disabled}
-                placeholder="Bank Address"
-                {...form2.getInputProps("bankAddress")}
-                errorProps={{
-                  fz: 12,
-                }}
-              />
-            </Flex>
+              {processing && <Loader type="oval" size={24} color="#12B76A" />}
+            </Group>
+          )}
 
-            <Flex gap={20} mt={24}>
-              <Select
-                searchable
-                placeholder="Select Country"
-                classNames={{ input: styles.input, label: styles.label }}
-                flex={1}
-                label="Country"
-                data={countries}
-                disabled={disabled}
-                {...form2.getInputProps("destinationCountry")}
-              />
-            </Flex>
+          {validated && (
+            <>
+              <Flex gap={20} mt={24}>
+                <TextInput
+                  placeholder="Enter Bank Name"
+                  classNames={{ input: styles.input, label: styles.label }}
+                  flex={1}
+                  size="lg"
+                  label={
+                    <Text fz={14} c="#667085">
+                      Bank
+                    </Text>
+                  }
+                  disabled={disabled}
+                  {...form2.getInputProps("destinationBank")}
+                  errorProps={{
+                    fz: 12,
+                  }}
+                />
+              </Flex>
 
-            <Flex gap={20} mt={24}>
-              <NumberInput
-                flex={1}
-                classNames={{ input: styles.input, label: styles.label }}
-                description={
-                  <Text fz={12}>
-                    {Number(form2.values.amount) >
-                    Number(account?.accountBalance)
-                      ? `Insufficient Balance`
-                      : ""}
-                  </Text>
-                }
-                styles={{
-                  description: {
-                    color: "var(--prune-warning)",
-                  },
-                  input: {
-                    border:
-                      Number(form2.values.amount) >
+              <Flex gap={20} mt={24}>
+                <TextInput
+                  classNames={{ input: styles.input, label: styles.label }}
+                  flex={1}
+                  size="lg"
+                  label={
+                    <Text fz={14} c="#667085">
+                      Bank Address
+                    </Text>
+                  }
+                  disabled={disabled}
+                  placeholder="Bank Address"
+                  {...form2.getInputProps("bankAddress")}
+                  errorProps={{
+                    fz: 12,
+                  }}
+                />
+              </Flex>
+
+              <Flex gap={20} mt={24}>
+                <Select
+                  searchable
+                  placeholder="Select Country"
+                  classNames={{ input: styles.input, label: styles.label }}
+                  flex={1}
+                  label="Country"
+                  data={countries}
+                  disabled={disabled}
+                  {...form2.getInputProps("destinationCountry")}
+                />
+              </Flex>
+
+              <Flex gap={20} mt={24}>
+                <NumberInput
+                  flex={1}
+                  classNames={{ input: styles.input, label: styles.label }}
+                  description={
+                    <Text fz={12}>
+                      {Number(form2.values.amount) >
                       Number(account?.accountBalance)
-                        ? "1px solid red"
-                        : "1px solid #eaecf0",
-                  },
-                }}
-                label={
-                  <Text fz={14} c="#667085">
-                    Amount
-                  </Text>
-                }
-                hideControls
-                size="lg"
-                placeholder="Enter amount"
-                {...form2.getInputProps("amount")}
-                errorProps={{
-                  fz: 12,
-                }}
-              />
-            </Flex>
+                        ? `Insufficient Balance`
+                        : ""}
+                    </Text>
+                  }
+                  styles={{
+                    description: {
+                      color: "var(--prune-warning)",
+                    },
+                    input: {
+                      border:
+                        Number(form2.values.amount) >
+                        Number(account?.accountBalance)
+                          ? "1px solid red"
+                          : "1px solid #eaecf0",
+                    },
+                  }}
+                  label={
+                    <Text fz={14} c="#667085">
+                      Amount
+                    </Text>
+                  }
+                  hideControls
+                  size="lg"
+                  placeholder="Enter amount"
+                  {...form2.getInputProps("amount")}
+                  errorProps={{
+                    fz: 12,
+                  }}
+                />
+              </Flex>
 
-            <Flex gap={20} mt={24} direction="column">
-              <Text fz={14} c="#667085" m={0} p={0}>
-                Upload supporting document (Optional)
-              </Text>
-              <DropzoneComponent<typeof sendMoneyRequest>
-                style={{ flex: 1 }}
-                otherForm={form2}
-                formKey="invoice"
-                uploadedFileUrl={form2.values.invoice}
-                isUser
-              />
-            </Flex>
+              <Flex gap={20} mt={24} direction="column">
+                <Text fz={14} c="#667085" m={0} p={0}>
+                  Upload supporting document (Optional)
+                </Text>
+                <DropzoneComponent<typeof sendMoneyRequest>
+                  style={{ flex: 1 }}
+                  otherForm={form2}
+                  formKey="invoice"
+                  uploadedFileUrl={form2.values.invoice}
+                  isUser
+                />
+              </Flex>
 
-            <Flex gap={20} mt={24}>
-              <Textarea
-                flex={1}
-                autosize
-                minRows={3}
-                size="lg"
-                classNames={{ input: styles.textarea, label: styles.label }}
-                label={
-                  <Text fz={14} c="#667085">
-                    Narration
-                  </Text>
-                }
-                placeholder="Enter narration"
-                {...form2.getInputProps("narration")}
-                errorProps={{
-                  fz: 12,
-                }}
-              />
-            </Flex>
-          </>
-        )}
+              <Flex gap={20} mt={24}>
+                <Textarea
+                  flex={1}
+                  autosize
+                  minRows={3}
+                  size="lg"
+                  classNames={{ input: styles.textarea, label: styles.label }}
+                  label={
+                    <Text fz={14} c="#667085">
+                      Narration
+                    </Text>
+                  }
+                  placeholder="Enter narration"
+                  {...form2.getInputProps("narration")}
+                  errorProps={{
+                    fz: 12,
+                  }}
+                />
+              </Flex>
+            </>
+          )}
+        </ScrollArea>
 
         <Flex mt={24} justify="flex-end" gap={15}>
           <SecondaryBtn
